Fix rank errors when there dropping between two stories with only 1 rank seperating them.
The updateRank logic assumed that rank was an integer and that Rally
balanced the ranks to be sure no two stories had ranks right next to
each other. Some experimentation proved that Rally make sure that a
rank change leaves the maximum room possible between ranks but will not
touch surrounding rankings. Also, the rank field is not an integer and
can support decimals. Therefore the better ranking stragegy is to set
the new rank to the average of the two ranks it is being dropped
between.