Fix ranking refresh
When changing the rank of a story the rank got changed in Rally but we
did not update the data in the current view. Therefore, if another rank
update request was made that involved the same story then it was not
properly updated. The fix was to update the model and view with the
stories new rank.