Explicitly use backlogElement
On the main page I had two variables named backlog one to refer to the
object and the other to review to the HTML element that contained the
backlog. In a previous commit I added Element to all the HTML element
variables I used but missed adding it in this one place. This fixes the
previous bug.