Test getting seurity token from Rally
After adding a test harness to my app I now want to start using it. The
first functional test I want to add is to test manipulating stories. In
order to create a test story in Rally I will need to get a security
key. This commit tests the retrieval of said key.